chore: 업로드 컴포넌트화
